Major Version update Further cleanup of exclusion and cleanup code Improvements to launch mode handling with original injection method loading filenames relative to app Experimental settings checkbox removed for simplicity, some configurations available through launch option Update to Net 8.0, no breaking API changes required, should compile to both
Early new feature:
Implement mutex and named pipe server with client itself, separate from shared memory with the backend. Currently it will prevent opening multiple frontend instances (not backwards compatible) and dragging a zip file onto the exe or starting it with a path to a zipfile as the launch arg will auto import the config in the running instance. Next step will be to implement a URI redirector/interceptor to handle downloads from discord